refactor(connections): rotate peer card chevron and relocate NetworkInfoDto tests to commonTest
